+++ packages/react/src/components/form/form.tsx
@@ -14,42 +14,134 @@ import {
   ConfigurationOverrides,
   ConfigurationOverridesProps,
 } from "../configuration-overrides";
-import { LoginOptions } from "../../domain/types";
+import { Handle, HandleType, LoginOptions } from "../../domain/types";
+import React, { useCallback } from "react";
+import { Slots, useSlots } from "../slot";
+import { Factor } from "@slashid/slashid";
+import { Flag } from "../input";
 
 export type Props = ConfigurationOverridesProps & {
   className?: string;
   onSuccess?: CreateFlowOptions["onSuccess"];
   middleware?: LoginOptions["middleware"];
+  children?: Slots<
+    "initial" | "authenticating" | "success" | "error" | "footer"
+  >; // TS does not enforce this, but it is used for documentation
 };
 
-export const Form: React.FC<Props> = ({
+type PayloadOptions = {
+  handleType?: HandleType;
+  handleValue?: string;
+  flag?: Flag;
+};
+
+export type InternalFormContextType = {
+  flowState: ReturnType<typeof useFlowState> | null;
+  lastHandle?: Handle;
+  submitPayloadRef: React.MutableRefObject<PayloadOptions>;
+  handleSubmit: (factor: Factor, handle?: Handle) => void;
+  selectedFactor?: Factor;
+  setSelectedFactor: React.Dispatch<React.SetStateAction<Factor | undefined>>;
+};
+export const InternalFormContext = React.createContext<InternalFormContextType>(
+  {
+    flowState: null,
+    lastHandle: undefined,
+    submitPayloadRef: { current: {} },
+    handleSubmit: () => null,
+    selectedFactor: undefined,
+    setSelectedFactor: () => null,
+  }
+);
+export const useInternalFormContext = () => {
+  return React.useContext(InternalFormContext);
+};
+
+/**
+ * Render a form that can be used to sign in or sign up a user.
+ * The form can be customized significantly using the built-in slots and CSS custom properties.
+ * Check the documentation for more information.
+ */
+export const Form = ({
   className,
   onSuccess,
   factors,
   text,
-  middleware
-}) => {
+  middleware,
+  children,
+}: Props) => {
   const flowState = useFlowState({ onSuccess });
   const { showBanner } = useConfiguration();
   const { lastHandle } = useLastHandle();
+  const submitPayloadRef = React.useRef<PayloadOptions>({
+    handleType: undefined,
+    handleValue: undefined,
+    flag: undefined,
+  });
+  const [selectedFactor, setSelectedFactor] = React.useState<
+    Factor | undefined
+  >();
+  const { status } = flowState;
+
+  const defaultSlots = React.useMemo(() => {
+    const slots = {
+      footer: showBanner ? <Footer /> : null,
+      initial: status === "initial" ? <Initial /> : undefined,
+      authenticating:
+        status === "authenticating" ? (
+          <Authenticating flowState={flowState} />
+        ) : undefined,
+      success:
+        status === "success" ? <Success flowState={flowState} /> : undefined,
+      error: status === "error" ? <Error flowState={flowState} /> : undefined,
+    };
+
+    return slots;
+  }, [status, showBanner, flowState]);
+
+  const slots = useSlots({ children, defaultSlots });
+
+  const handleSubmit = useCallback(
+    (factor: Factor, handle?: Handle) => {
+      if (flowState.status !== "initial") return;
+
+      flowState.logIn(
+        {
+          factor,
+          handle,
+        },
+        { middleware }
+      );
+    },
+    [flowState, middleware]
+  );
 
   return (
-    <div className={clsx("sid-form", styles.form, className)}>
-      <ConfigurationOverrides text={text} factors={factors}>
-        {flowState.status === "initial" && (
-          <FormProvider>
-            <Initial flowState={flowState} lastHandle={lastHandle} middleware={middleware} />
-          </FormProvider>
-        )}
-        {flowState.status === "authenticating" && (
-          <FormProvider>
-            <Authenticating flowState={flowState} />
-          </FormProvider>
-        )}
-        {flowState.status === "error" && <Error flowState={flowState} />}
-        {flowState.status === "success" && <Success flowState={flowState} />}
-        {showBanner ? <Footer /> : null}
-      </ConfigurationOverrides>
-    </div>
+    <InternalFormContext.Provider
+      value={{
+        flowState,
+        lastHandle,
+        handleSubmit,
+        submitPayloadRef,
+        selectedFactor,
+        setSelectedFactor,
+      }}
+    >
+      <div className={clsx("sid-form", styles.form, className)}>
+        <ConfigurationOverrides text={text} factors={factors}>
+          {flowState.status === "initial" && (
+            <FormProvider>{slots.initial}</FormProvider>
+          )}
+          {flowState.status === "authenticating" && (
+            <FormProvider>{slots.authenticating}</FormProvider>
+          )}
+          {flowState.status === "success" && slots.success}
+          {flowState.status === "error" && slots.error}
+          {slots.footer}
+        </ConfigurationOverrides>
+      </div>
+    </InternalFormContext.Provider>
   );
 };
+
+Form.Initial = Initial;