+++ packages/react/src/main.ts
@@ -1,5 +1,6 @@
 import { DynamicFlow } from "./components/dynamic-flow";
 import { Form } from "./components/form";
+import { Slot } from "./components/slot";
 import { GDPRConsentDialog } from "./components/gdpr-consent-dialog";
 import { Groups } from "./components/groups";
 import { LoggedIn } from "./components/logged-in";
@@ -36,6 +37,7 @@ export {
   LoggedOut,
   MultiFactorAuth,
   OrganizationSwitcher,
+  Slot,
   SlashIDContext,
   SlashIDLoaded,
   SlashIDProvider,